+++ src/components/features/LetterWrite/InputRecipient/Forms/RelationshipForm.tsx
@@ -1,20 +1,14 @@
-import BottomButton from "@/src/components/common/Buttons/BottomButton";
 import { letterWriteInputState } from "@/src/store/LetterWrite";
-import { getNavigateNextLink } from "@/src/utils/helper/LetterWrite";
-import { useRouter } from "next/router";
 import { ReactElement } from "react";
 import { useRecoilValue } from "recoil";
 import { LetterWriteTagContainer } from "../../common";
+import { useBottomButton } from "../Hooks";
 import * as S from "../styled";
 
 const RelationshipForm = (): ReactElement => {
-  const router = useRouter();
-  const { type } = router.query;
-  const nextLink = getNavigateNextLink(type);
   const letterWriteInputObjectState = useRecoilValue(letterWriteInputState);
-  const onClickNext = () => {
-    router.push(nextLink);
-  };
+  const bottomButton = useBottomButton;
+
   return (
     <>
       <S.LetterWriteH1>어떤 관계인가요?</S.LetterWriteH1>
@@ -25,15 +19,7 @@ const RelationshipForm = (): ReactElement => {
         </S.LetterWriteBody1>
       </S.RelationShipReceiverContainer>
       <LetterWriteTagContainer />
-      <S.BottomButtonContainer>
-        <BottomButton
-          name="다음"
-          isDark={true}
-          isRound={true}
-          disabled={!letterWriteInputObjectState.relationship}
-          onClick={onClickNext}
-        />
-      </S.BottomButtonContainer>
+      {bottomButton}
     </>
   );
 };
