+++ src/components/features/LetterWrite/InputRecipient/Hooks/index.ts
@@ -1,0 +1,3 @@
+import useBottomButton from "./useBottomButton";
+
+export { useBottomButton };
